# docs/api-audit/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tecbunny API Surface",
    "version": "1.0.0",
    "description": "Generated from repository source by scripts/audit-api-surface.mjs."
  },
  "servers": [
    {
      "url": "{{API_BASE_URL}}"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/admin/products": {
      "get": {
        "description": "GET /api/admin/products",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Observed or expected 401 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [],
        "x-route-file": "apps/mgmt/src/app/api/admin/products/route.ts",
        "x-database-tables": [
          "information_schema.columns",
          "products"
        ],
        "x-frontend-usage": [
          "apps/superadmin/src/app/superadmin/mgmt/catalogue/page.tsx",
          "packages/admin-ui/src/components/admin-homepage-settings.tsx",
          "packages/admin-ui/src/components/admin-pricing.tsx",
          "packages/admin-ui/src/components/admin-products-new.tsx"
        ]
      }
    },
...
    "/api/roles": {
      "get": {
        "description": "GET /api/roles",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/superadmin/src/app/api/roles/route.ts",
        "x-database-tables": [],
        "x-frontend-usage": []
      },
      "post": {
        "description": "POST /api/roles",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/superadmin/src/app/api/roles/route.ts",
        "x-database-tables": [],
        "x-frontend-usage": []
      }
    },
...
    "/api/services/tickets": {
      "get": {
        "description": "GET /api/services/tickets",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/services/tickets/route.ts",
        "x-database-tables": [],
        "x-frontend-usage": []
      },
      "post": {
        "description": "POST /api/services/tickets",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/services/tickets/route.ts",
        "x-database-tables": [],
        "x-frontend-usage": []
      }
    },
    "/api/services/tickets/{id}": {
      "get": {
        "description": "GET /api/services/tickets/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/services/tickets/[id]/route.ts",
        "x-database-tables": [
          "service_tickets"
        ],
        "x-frontend-usage": []
      },
      "put": {
        "description": "PUT /api/services/tickets/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/api/src/app/api/services/tickets/[id]/route.ts",
        "x-database-tables": [
          "service_tickets"
        ],
        "x-frontend-usage": []
      }
    },
...
  }
}